relocate: Mypage 폴더 안에 있던 assets를 바깥으로 꺼냄
