+++ src/pages/List/components/IdolDetail.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import "./IdolDetail.css";
-import checkIcon from "../../Mypage/assets/images/ic_check.svg";
+import checkIcon from "../../../assets/images/ic_check.svg";
 
 function IdolDetail({
   idolData,
