+++ src/pages/Mypage/components/FavoriteIdolList.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import "./FavoriteIdolList.css";
-import useSaveMyIdol from "../hooks/useSaveMyIdol";
+import useSaveMyIdol from "../../../hooks/useSaveMyIdol";
 
 import MyFavoriteIdolList from "./MyFavoriteIdolList/MyFavoriteIdolList";
 import AddFavoriteIdolList from "./AddFavoriteIdolList/AddFavoriteIdolList";
